+++ pyclip/__init__.py
