+++ pyclip/__main__.py
