+++ pyclip/base.py
