+++ pyclip/macos_clip.py
