+++ pyclip/util.py
