+++ pyclip/win_clip.py
