+++ tests/test_clipboard.py
@@ -1,10 +1,10 @@
 import sys, os
 try:
-    import pyperclip3 as clip
+    import pyclip as clip
 except ImportError:
     # XXX: probably shouldn't do this, but oh well ¯\_(ツ)_/¯
     sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-    import pyperclip3 as clip
+    import pyclip as clip
 import pytest
 
 from unittest import mock
@@ -39,7 +39,7 @@ def test_clear():
 
 @pytest.mark.skipif(sys.platform != 'darwin', reason='This test is for MacOS only')
 def test_copypaste_pbcopy_backend():
-    from pyperclip3.macos_clip import _PBCopyPBPasteBackend, MacOSClip
+    from pyclip.macos_clip import _PBCopyPBPasteBackend, MacOSClip
     backend = _PBCopyPBPasteBackend()
     clip = MacOSClip(_backend=backend)
     clip.copy('foo')
@@ -48,7 +48,7 @@ def test_copypaste_pbcopy_backend():
     assert clip.paste(text=True) == 'foo'
 @pytest.mark.skipif(sys.platform != 'darwin', reason='This test is for MacOS only')
 def test_copypaste_unicode_pbcopy_backend():
-    from pyperclip3.macos_clip import _PBCopyPBPasteBackend, MacOSClip
+    from pyclip.macos_clip import _PBCopyPBPasteBackend, MacOSClip
     backend = _PBCopyPBPasteBackend()
     clip = MacOSClip(_backend=backend)
     unicode = 'א ב ג ד ה ו ז ח ט י ך כ ל ם מ ן נ ס ע ף פ ץ צ ק ר ש ת װ ױ'
@@ -58,7 +58,7 @@ def test_copypaste_unicode_pbcopy_backend():
 
 @pytest.mark.skipif(sys.platform != 'darwin', reason='This test is for MacOS only')
 def test_clear_pbcopy_backend():
-    from pyperclip3.macos_clip import _PBCopyPBPasteBackend, MacOSClip
+    from pyclip.macos_clip import _PBCopyPBPasteBackend, MacOSClip
     backend = _PBCopyPBPasteBackend()
     clip = MacOSClip(_backend=backend)
     clip.copy('foo')
@@ -68,8 +68,8 @@ def test_clear_pbcopy_backend():
     assert not data, f'clipboard contents unexpectly present: {repr(data)}'
 
 def test_cli():
-    from pyperclip3.cli import _main, main
-    args = ['pyperclip3', 'copy']
+    from pyclip.cli import _main, main
+    args = ['pyclip', 'copy']
     import io
     stdin = io.BytesIO(b"foo")
     class MockStdin:
@@ -78,7 +78,7 @@ def test_cli():
     with mock.patch('sys.exit', new=lambda x: x):
         with mock.patch('sys.argv', new=args), mock.patch('sys.stdin', new=MockStdin(b'foo')):
             main()
-        args = ['pyperclip3', 'paste']
+        args = ['pyclip', 'paste']
         class MockStdout:
             def __init__(self):
                 self.buffer = io.BytesIO()
@@ -86,7 +86,7 @@ def test_cli():
         with mock.patch('sys.argv', new=args), mock.patch('sys.stdout', new=stdout):
             main()
         assert stdout.buffer.getvalue() == b'foo'
-        args = ['pyperclip3', 'clear']
+        args = ['pyclip', 'clear']
         with mock.patch('sys.argv', new=args):
             main()
         assert not clip.paste()
